extract integration test
